requirements.txt
asgiref==3.8.1
Django==4.2.13
django-ckeditor==6.7.1
django-js-asset==2.2.0
mysql-connector-python==9.0.0
sqlparse==0.5.0
typing_extensions==4.12.2
